requirements.txt
torch
transformers
pillow
flask
requests
gradio
